Performance and cosmetic improvements
Fixed performance and memory usage issue calculating average price for a commodity on systems with a large number of imports/exports.

Added new loader when fetching data for a  system or commodity not in the cache.

Fixed a number of small CSS alignment / spacing issues.

Refactored layout for CSS files to reduce nesting.

Add basic metadata to pages.