Improve system information summary
* Add count and list of known stations
* Add count and list of known carriers
* Add last updated timestamp

Also:

* Improve appearance of progress bars (e.g. on commodity page)
* Improve default styling for lists
* Tweaks to font sie of tabs on mobile
* Fix bug in muted text CSS variable

This is intended as an interim implementation which fetches the list of known markets.

It will be updated to use canonical and more complete index of station data when the API endpoints for that data are published.

At that point the UI should also be updated to make it easier to browse and view other data (e.g. economy type, faction, services, etc) for each location.