Refactor how commodity filter options are persisted
Removed persisting options via localStorage and moved it to URL state.

This is more ephemeral, but is a less confusing user experience, and still allows links to views to be easily shared.

It's far from perfect and I expect to refine it further - and the render and network behaviour needs optimisation.