Optimise to reduce number of API requests
Minor, slightly hacky, change that signficiantly reduces the number of  network requests made by avoiding unintended duplicate requests for the same data.

This "should" be resolved by refactoring the component logic and cleaning up the code, but that's a lot of work for little to no performance gain, and would probably introduce new bugs (until they were squashed) and there are more intersting features to work on.