Add support for new commodities
They are automatically added and tracked by the database but they are not automatically surfaced in the website.

This should probably be refactored so that the commodiites file is served by the backend so I don't have to manually update this file when new commodities are added to the game (even if by default they names are not nicely formatted and they are missing descriptions).