New autocomplete control for system names
This is still very hacky, as it touches a lot of behaviour and I'm trying to preserve much of the former behaviour (bugs aside).

I want to keep working on this and radically simplify it, but I think it's working decently well at this point where I can push the change live and come back to it to refine it.

It still has some undesirable behaviour but overall I think the UX is an improvement over the currently live implementation.