Refactor how highlighted items are displayed
Instead of using a contrast filter now shifts the hue of the current theme.

This is a analogous color rather than a contrasting one, but enough to be visually distinct.

I have some concerns about the contrast levels being too for some base colors, this might be something that can be address in a future update or it might be that only some color options are going to work well for those that need high contrast.

I ran into some browser limitations with Firefox but side stepped them for now; Firefox does not yet support css if() syntax AND does not gracefully fallback, it might need special case handling in future.